move server hooks to a hopefully better place
